requirements.txt
fastapi
uvicorn
requests
sentence-transformers
scikit-learn
numpy
python-dotenv